Modify and create router page
