+++ audience-check-in/frontend/audience-check-in-web/src/pages/AudiencesPage.vue
@@ -1,0 +1,5 @@
+<template>
+    <h5 class="q-pl-xl">
+        Audiences
+    </h5>
+</template>
