+++ audience-check-in/frontend/audience-check-in-web/src/router/routes.ts
@@ -4,7 +4,10 @@ const routes: RouteRecordRaw[] = [
   {
     path: '/',
     component: () => import('layouts/MainLayout.vue'),
-    children: [{ path: '', component: () => import('pages/IndexPage.vue') }],
+    children: [
+      { path: '', component: () => import('pages/HomePage.vue') },
+      { path: 'audiences', component: () => import('pages/AudiencesPage.vue') },
+    ],
   },
 
   // Always leave this as last one,
